+++ fuse_kernel.go
@@ -546,7 +546,7 @@ type releaseIn struct {
 	Fh           uint64
 	Flags        uint32
 	ReleaseFlags uint32
-	LockOwner    uint32
+	LockOwner    uint64
 }
 
 type flushIn struct {
